common calendar bits
